+++ changelog.md
@@ -2,5 +2,7 @@
 
 ## Hide Eject
 
+[v0.3] Changed how registry values are modified  
+[v0.3] Restore Windows File Explorer windows after restarting the process.  
 [v0.2] Added the `-Permanent` switch, which creates a scheduled task to automatically reapply the patch on every restart.  
 [v0.1] Initial version